requirements.txt
scrapy
csvs-to-sqlite
sqlite-utils
python-dateutil
